// src/api/models/DealListResponse.ts
export interface Data {
  Deals: Deal[]
}

export interface Deal {
  Deal_No: string
  Deal_Date: string
  Comp_FName: string
  Branch_ID: string
  Branch_SName: string
  Branch_Code: string
  Branch_Address1: string
  Branch_Address2: string
  Branch_City: string
  Branch_Country: string
  Branch_Tel1: string
  Branch_Fax1: string
  Branch_Email1: string
  Delivery_Branch_ID: string
  Delivery_Branch_Code: string
  Delivery_Branch_SName: string
  EX_Company: string
  StlD: string
  Party_Deal_Remarks: string
  SN_Name: string
  SN_Address1: string
  SN_City: string
  SN_Country: string
  SN_Tel_Mobile: string
  SN_Relationship: string
  SN_Purpose: string
  CTypeD: string
  CCType: string
  DetailCode: string
  DetailCodeDescription: string
  TrLevel: string
  TrLevelD: string
  ROTypeID: string
  ROTypeDD: string
  Deal_Trlevel: string
  Deal_TrlevelD: string
  Delivery_At: string
  PayIN_CurrID: string
  PayIN_Amt: string
  PayIN_PKR: string
  PayIN_Rate: string
  PayOut_CurrID: string
  PayOut_Rate: string
  PayOut_Amt: string
  PayOut_PKR: string
  Party_Profile: string
  Party_Code: string
  Party_Title: string
  Party_Name: string
  Party_CStatus_ID: string
  Party_CStatus_D: string
  Party_ITypeID: string
  Party_IType: string
  Party_ITypeRef: string
  Party_IType_Expiry: string
  Party_Tel_Mobile: string
  Party_Email: string
  Party_Address1: string
  Party_City: string
  Party_Country: string
  AccActivity_1_ID: string
  AccActivity_1_D: string
  AccActivity_1_DD: string
  U_S: string
  PDate: string
}